🌱 refactor: split oversized drilldown components into hooks and parts
Each drilldown now has:
- use{Name}DrillDown.ts: Data fetching and state management hook
- {Name}DrillDown.parts.tsx: Reusable UI sub-components
- {Name}DrillDown.tsx: Main component (presentational)

Results (lines / unique hooks):
- PolicyDrillDown: 553 → 328, 15 → 7
- PVCDrillDown: 414 → 252, 12 → 5
- ReplicaSetDrillDown: 435 → 310, 12 → 6
- EventsDrillDown: 458 → 251, 13 → 7
- NamespaceDrillDown: 443 → 330, 10 → 7

All files now meet acceptance criteria: <350 lines AND <10 hook calls.
